==> Major Producers workspace/CODE/1r_extract_agreste.py <==
#%% About
'''
This reads raw data from Agreste, France's national agriculture database.
'''
#%% Import

# os.path.join(RAWDATA_FOLDER ,'agreste_animalfeed_productionandvalue.csv')
# os.path.join(RAWDATA_FOLDER ,'agreste_livestockslaughter_2011_2022.csv')
